Upgrade jest version to latest and move tests into __tests__ directory structure
