+++ .eslintignore
@@ -5,3 +5,4 @@ locale
 preset-react-app-custom
 build
 **/*.ts
+**/__tests__/**
