+++ src/parser/__tests__/Root.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import Root from './Root'
+import Root from '../../Root'
 
 describe('Root', () => {
 	it('renders without crashing', () => {
